+++ tests/helpers/helpers.py
@@ -1,35 +1,19 @@
-# The MIT License (MIT)
-# Copyright © 2024 Opentensor Foundation
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-# documentation files (the “Software”), to deal in the Software without restriction, including without limitation
-# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
-# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all copies or substantial portions of
-# the Software.
-#
-# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
-# THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-# DEALINGS IN THE SOFTWARE.
-
-from collections import deque
+import asyncio
 import json
+import time
+from collections import deque
 from typing import Union
-
-from websockets.sync.client import ClientConnection, ClientProtocol
-from websockets.uri import parse_uri
 
 from bittensor_wallet.mock.wallet_mock import MockWallet as _MockWallet
 from bittensor_wallet.mock.wallet_mock import get_mock_coldkey
 from bittensor_wallet.mock.wallet_mock import get_mock_hotkey
 from bittensor_wallet.mock.wallet_mock import get_mock_wallet
+from websockets.asyncio.client import ClientConnection, ClientProtocol
+from websockets.uri import parse_uri
 
-from bittensor.utils.balance import Balance
 from bittensor.core.chain_data import AxonInfo, NeuronInfo, PrometheusInfo
-from tests.helpers.integration_websocket_data import WEBSOCKET_RESPONSES
+from bittensor.utils.balance import Balance
+from tests.helpers.integration_websocket_data import WEBSOCKET_RESPONSES, METADATA
 
 
 def __mock_wallet_factory__(*_, **__) -> _MockWallet:
@@ -129,9 +113,10 @@ class FakeWebsocket(ClientConnection):
 
     def __init__(self, *args, seed, **kwargs):
         protocol = ClientProtocol(parse_uri("ws://127.0.0.1:9945"))
-        super().__init__(socket=None, protocol=protocol, **kwargs)
+        super().__init__(protocol=protocol, **kwargs)
         self.seed = seed
         self.received = deque()
+        self._lock = asyncio.Lock()
 
     def send(self, payload: str, *args, **kwargs):
         received = json.loads(payload)
@@ -139,16 +124,32 @@ class FakeWebsocket(ClientConnection):
         self.received.append((received, id_))
 
     def recv(self, *args, **kwargs):
+        while len(self.received) == 0:
+            time.sleep(0.1)
         item, _id = self.received.pop()
         try:
-            response = WEBSOCKET_RESPONSES[self.seed][item["method"]][
-                json.dumps(item["params"])
-            ]
-            response["id"] = _id
+            if item["method"] == "state_getMetadata":
+                response = {"jsonrpc": "2.0", "id": _id, "result": METADATA}
+            else:
+                response = WEBSOCKET_RESPONSES[self.seed][item["method"]][
+                    json.dumps(item["params"])
+                ]
+                response["id"] = _id
             return json.dumps(response)
         except (KeyError, TypeError):
             print("ERROR", self.seed, item["method"], item["params"])
             raise
 
     def close(self, *args, **kwargs):
+        pass
+
+
+class FakeConnectContextManager:
+    def __init__(self, seed):
+        self.seed = seed
+
+    def __enter__(self):
+        return FakeWebsocket(seed=self.seed)
+
+    def __exit__(self, exc_type, exc, tb):
         pass