+++ tests/integration_tests/test_metagraph_integration.py
@@ -1,19 +1,4 @@
-# The MIT License (MIT)
-# Copyright © 2024 Opentensor Foundation
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-# documentation files (the “Software”), to deal in the Software without restriction, including without limitation
-# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
-# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all copies or substantial portions of
-# the Software.
-#
-# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
-# THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-# DEALINGS IN THE SOFTWARE.
+from unittest import mock
 
 import bittensor
 import torch
@@ -39,22 +24,41 @@ class TestMetagraph:
         print(self.metagraph)
 
     def test_lite_sync(self):
-        self.metagraph.sync(lite=True, subtensor=self.sub)
+        with mock.patch.object(
+            self.sub, "get_metagraph_info", return_value=mock.MagicMock()
+        ):
+            self.metagraph.sync(lite=True, subtensor=self.sub)
 
     def test_full_sync(self):
-        self.metagraph.sync(lite=False, subtensor=self.sub)
+        with mock.patch.object(
+            self.sub, "get_metagraph_info", return_value=mock.MagicMock()
+        ):
+            self.metagraph.sync(lite=False, subtensor=self.sub)
 
     def test_sync_block_0(self):
-        self.metagraph.sync(lite=True, block=0, subtensor=self.sub)
+        with mock.patch.object(
+            self.sub, "get_metagraph_info", return_value=mock.MagicMock()
+        ):
+            self.metagraph.sync(lite=True, block=0, subtensor=self.sub)
 
     def test_load_sync_save(self):
-        self.metagraph.sync(lite=True, subtensor=self.sub)
-        self.metagraph.save()
-        self.metagraph.load()
-        self.metagraph.save()
+        with mock.patch.object(
+            self.sub, "neurons_lite", return_value=[]
+        ), mock.patch.object(
+            self.sub, "get_metagraph_info", return_value=mock.MagicMock()
+        ):
+            self.metagraph.sync(lite=True, subtensor=self.sub)
+            self.metagraph.save()
+            self.metagraph.load()
+            self.metagraph.save()
 
     def test_load_sync_save_from_torch(self):
-        self.metagraph.sync(lite=True, subtensor=self.sub)
+        with mock.patch.object(
+            self.sub, "neurons_lite", return_value=[]
+        ), mock.patch.object(
+            self.sub, "get_metagraph_info", return_value=mock.MagicMock()
+        ):
+            self.metagraph.sync(lite=True, subtensor=self.sub)
 
         def deprecated_save_torch(metagraph):
             save_directory = get_save_dir(metagraph.network, metagraph.netuid)
@@ -77,7 +81,6 @@ class TestMetagraph:
         assert "n" in state
         assert "block" in state
         assert "stake" in state
-        assert "total_stake" in state
         assert "ranks" in state
         assert "trust" in state
         assert "consensus" in state