+++ tests/unit_tests/extrinsics/asyncex/test_registration.py
@@ -2,7 +2,7 @@ import pytest
 from bittensor_wallet import Wallet
 
 from bittensor.core import async_subtensor
-from bittensor.core.extrinsics import async_registration
+from bittensor.core.extrinsics.asyncex import registration as async_registration
 
 
 @pytest.fixture(autouse=True)
@@ -69,12 +69,11 @@ async def test_do_pow_register_success(subtensor, mocker):
     subtensor.substrate.create_signed_extrinsic.asseert_awaited_once_with(
         call=fake_call, keypair=fake_wallet.hotkey
     )
-    subtensor.substrate.submit_extrinsic.asseert_awaited_once_with(
+    subtensor.substrate.submit_extrinsic.assert_awaited_once_with(
         fake_extrinsic, wait_for_inclusion=True, wait_for_finalization=True
     )
-    fake_response.process_events.assert_called_once()
     assert result is True
-    assert error_message is None
+    assert error_message == ""
 
 
 @pytest.mark.asyncio
@@ -106,7 +105,7 @@ async def test_do_pow_register_failure(subtensor, mocker):
         subtensor.substrate, "submit_extrinsic", return_value=fake_response
     )
     mocked_format_error_message = mocker.patch.object(
-        async_registration, "format_error_message"
+        async_subtensor, "format_error_message"
     )
 
     # Call
@@ -125,10 +124,10 @@ async def test_do_pow_register_failure(subtensor, mocker):
         call=fake_call, keypair=fake_wallet.hotkey
     )
     subtensor.substrate.submit_extrinsic.asseert_awaited_once_with(
-        extrinsic=fake_extrinsic, wait_for_inclusion=True, wait_for_finalization=True
+        fake_extrinsic, wait_for_inclusion=True, wait_for_finalization=True
     )
 
-    mocked_format_error_message.assert_called_once_with(error_message=fake_err_message)
+    mocked_format_error_message.assert_called_once_with(fake_err_message)
     assert result_error_message == (False, mocked_format_error_message.return_value)
 
 
@@ -174,7 +173,7 @@ async def test_do_pow_register_no_waiting(subtensor, mocker):
         fake_extrinsic, wait_for_inclusion=False, wait_for_finalization=False
     )
     assert result is True
-    assert error_message is None
+    assert error_message == ""
 
 
 @pytest.mark.asyncio
@@ -215,8 +214,15 @@ async def test_register_extrinsic_success(subtensor, mocker):
     )
 
     # Asserts
-    mocked_subnet_exists.assert_called_once_with(1)
-    mocked_get_neuron.assert_called_once_with(hotkey_ss58="hotkey_ss58", netuid=1)
+    mocked_subnet_exists.assert_called_once_with(
+        1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
+    mocked_get_neuron.assert_called_once_with(
+        hotkey_ss58="hotkey_ss58",
+        netuid=1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
     mocked_create_pow.assert_called_once()
     mocked_do_pow_register.assert_called_once()
     mocked_is_hotkey_registered.assert_called_once_with(
@@ -265,8 +271,15 @@ async def test_register_extrinsic_success_with_cuda(subtensor, mocker):
     )
 
     # Asserts
-    mocked_subnet_exists.assert_called_once_with(1)
-    mocked_get_neuron.assert_called_once_with(hotkey_ss58="hotkey_ss58", netuid=1)
+    mocked_subnet_exists.assert_called_once_with(
+        1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
+    mocked_get_neuron.assert_called_once_with(
+        hotkey_ss58="hotkey_ss58",
+        netuid=1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
     mocked_create_pow.assert_called_once()
     mocked_do_pow_register.assert_called_once()
     mocked_is_hotkey_registered.assert_called_once_with(
@@ -304,8 +317,15 @@ async def test_register_extrinsic_failed_with_cuda(subtensor, mocker):
     )
 
     # Asserts
-    mocked_subnet_exists.assert_called_once_with(1)
-    mocked_get_neuron.assert_called_once_with(hotkey_ss58="hotkey_ss58", netuid=1)
+    mocked_subnet_exists.assert_called_once_with(
+        1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
+    mocked_get_neuron.assert_called_once_with(
+        hotkey_ss58="hotkey_ss58",
+        netuid=1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
     assert result is False
 
 
@@ -327,7 +347,10 @@ async def test_register_extrinsic_subnet_not_exists(subtensor, mocker):
     )
 
     # Asserts
-    mocked_subnet_exists.assert_called_once_with(1)
+    mocked_subnet_exists.assert_called_once_with(
+        1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
     assert result is False
 
 
@@ -351,7 +374,9 @@ async def test_register_extrinsic_already_registered(subtensor, mocker):
 
     # Asserts
     mocked_get_neuron.assert_called_once_with(
-        hotkey_ss58=fake_wallet.hotkey.ss58_address, netuid=1
+        hotkey_ss58=fake_wallet.hotkey.ss58_address,
+        netuid=1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
     )
     assert result is True
 
@@ -401,8 +426,15 @@ async def test_register_extrinsic_max_attempts_reached(subtensor, mocker):
     )
 
     # Asserts
-    mocked_subnet_exists.assert_called_once_with(1)
-    mocked_get_neuron.assert_called_once_with(hotkey_ss58="hotkey_ss58", netuid=1)
+    mocked_subnet_exists.assert_called_once_with(
+        1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
+    mocked_get_neuron.assert_called_once_with(
+        hotkey_ss58="hotkey_ss58",
+        netuid=1,
+        block_hash=subtensor.substrate.get_chain_head.return_value,
+    )
     assert mocked_create_pow.call_count == 3
     assert mocked_do_pow_register.call_count == 3