+++ tests/unit_tests/test_chain_data.py
@@ -1,23 +1,7 @@
-# The MIT License (MIT)
-# Copyright © 2024 Opentensor Foundation
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-# documentation files (the “Software”), to deal in the Software without restriction, including without limitation
-# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
-# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all copies or substantial portions of
-# the Software.
-#
-# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
-# THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-# DEALINGS IN THE SOFTWARE.
-
 import pytest
 import torch
 
+from async_substrate_interface.utils import json
 from bittensor.core.chain_data import AxonInfo, DelegateInfo
 from bittensor.core.chain_data.utils import ChainDataType
 
@@ -119,7 +103,19 @@ def test_eq(other, expected, test_case):
                 hotkey="hot",
                 coldkey="cold",
             ),
-            '{"version": 1, "ip": "127.0.0.1", "port": 8080, "ip_type": 4, "hotkey": "hot", "coldkey": "cold", "protocol": 4, "placeholder1": 0, "placeholder2": 0}',
+            json.dumps(
+                {
+                    "version": 1,
+                    "ip": "127.0.0.1",
+                    "port": 8080,
+                    "ip_type": 4,
+                    "hotkey": "hot",
+                    "coldkey": "cold",
+                    "protocol": 4,
+                    "placeholder1": 0,
+                    "placeholder2": 0,
+                }
+            ),
             "ID_to_string",
         ),
     ],